Added algolia logo + separated front page and main app.
