+++ src/components/simple_searchbar.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import algoliasearch from 'algoliasearch/lite';
+import alglogo from "../images/algolia-white.svg";
 
 import { Link } from "react-router-dom";
 
@@ -53,16 +54,19 @@ class SimpleSearchBar extends React.Component {
 
   render() {
     return (
-      
-      <form className="form-inline d-flex content-center" onSubmit={this.handleSubmit}>
-
-        <input type="text" className="form-control flex-fill mr-0 mr-sm-2 mb-3 mb-sm-0" id="inputText" autoComplete="off" placeholder="ASK A QUESTION" value={this.state.value} onChange={this.handleChange} />
-        <div className="mx-auto">
-        <Link to={"/results/" + this.state.value}>
-        <input type="submit" disabled={!this.state.isEnabled} className="btn btn-primary mx-auto" value="Submit" />
-        </Link>
-        </div>
-      </form>
+      <div>
+          <a className="text-left alg-logo" href="https://www.algolia.com" class="ais-PoweredBy-link" aria-label="Algolia">
+            <img className="alg-logo content-left" src={alglogo} alt="alglogo"></img>
+          </a>
+        <form className="form-inline d-flex content-center text-center" onSubmit={this.handleSubmit}>
+          <input type="text" className="form-control flex-fill mr-0 mr-sm-2 mb-3 mb-sm-0" id="inputText" autoComplete="off" placeholder="ASK A QUESTION" value={this.state.value} onChange={this.handleChange} />
+          <div className="mx-auto">
+          <Link to={"/results/" + this.state.value}>
+            <input type="submit" disabled={!this.state.isEnabled} className="btn btn-primary mx-auto" value="Submit" />
+          </Link>
+          </div>
+        </form>
+      </div>
 
     );
   }
