+++ src/css/styles.css
@@ -34,13 +34,18 @@ a:hover {
   background: transparent;
   background-color: rgba(153,153,153,0)
 }
-html { 
+body, html {
+  height: 100%;
+}
+body { 
   background: url(../images/gv.jpg) no-repeat center center fixed; 
+  height:100%;
   -webkit-background-size: cover;
   -moz-background-size: cover;
   -o-background-size: cover;
   background-size: cover;
 }
+
 .brand-logo {
   width: 10%;
 }
@@ -270,4 +275,10 @@ html {
   }
 
   .ais-Hits { margin: 1em 0; font-family: 'Varela Round'; }
+}
+
+.alg-logo{
+  text-align: left;
+  width: 126px;
+  height: 18px;
 }
