+++ src/pages/frontpage.js
@@ -1,51 +1,23 @@
 import React from 'react';
-import { BrowserRouter as Router, Route, Switch, Link, useParams } from "react-router-dom";
-
-import {getQuestions} from '../hooks/web_hooks'
+import { BrowserRouter as Link } from "react-router-dom";
 
 //import SimpleSearchBar from '../components/simple_searchbar'
-import SearchBarText from '../components/searchbar_text'
-import questionList from '../components/questionList';
-import Nav from '../components/nav';
+import SearchBarText from '../components/searchbar_text';
 
-import profile from '../pages/profilePage';
-
-function App() {
-  getQuestions();
+class FrontPage extends React.Component {
+  render(){
   return (
-    <Router>
-
-      <div>
-        <link href="./bootstrap.min.css" rel="stylesheet"></link>
-        <link href="./font-awesome.min.css" rel="stylesheet"></link>
-        <link href="https://fonts.googleapis.com/css?family=Varela+Round&display=swap" rel="stylesheet"></link>
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/instantsearch.css@7.3.1/themes/reset-min.css" integrity="sha256-t2ATOGCtAIZNnzER679jwcFcKYfLlw01gli6F6oszk8=" crossOrigin="anonymous"></link>
-        <script src="https://polyfill.io/v3/polyfill.min.js?features=default%2CArray.prototype.find%2CArray.prototype.includes%2CPromise%2CObject.assign%2CObject.entries"></script>
-        
-        <Nav />
-        <Route path="/" exact component={frontPage} />
-        <Route path="/results/*" exact component={questionList}/>
-        <Route path="/profile" exact component={profile} />
-        <Switch>
-          <Route path="/results/:id"> <RoutResults/></Route>
-        </Switch>
-
-      </div>
-    </Router>
-  );
-}
-
-const frontPage = () => (
-  <div className="App container">
+  <div>
+    {/* <img className="app-bg" src={appbg} alt="bg"/> */}
     <section id="signup" className="signup-section mt-5 justify-content-center d-flex h-100">
-        <div className="mx-auto text-center jumbotron">
-          <h1 className="text-white mb-4">Studique</h1>
-          <h3 className="text-white mb-5">
+        <div className="mx-auto jumbotron">
+          <h1 className="text-white mb-4 text-center">Studique</h1>
+          <h3 className="text-white mb-5 text-center">
             Crowdsourcing The College Experience
           </h3>
           {/* <SimpleSearchBar/> */}
           <SearchBarText/>
-          <div className="text-white text-font">
+          <div className="text-white text-font text-center">
             <div className="my-2">
             Want to help others? 
             </div>
@@ -57,16 +29,7 @@ const frontPage = () => (
     </section>
   </div>
 );
-
-function RoutResults() {
-  // access dynamic URL variables
-  let { id } = useParams();
-
-  return (
-    <div>
-      <h3>ID: {id}</h3>
-    </div>
-  );
+  }
 }
 
-export default App;
+export default FrontPage;
